% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dhs_utils.R
\name{create_strata_dhs}
\alias{create_strata_dhs}
\title{Add strata column to DHS birth recode module}
\usage{
create_strata_dhs(df)
}
\arguments{
\item{df}{data.frame with DHS dataset as downloaded.}
}
\value{
data.frame with "strata" column added.
}
\description{
Add strata column for DHS birth recode manual, based on
combining existing columns, e.g. a column for region and a column for
rurality. This is based on the DHS .do file available at:
https://github.com/DHSProgram/DHS-Analysis-Code/blob/main/Intro_DHSdata_Analysis/2_SurveyDesign/Survey_strata.do
and is consistent with the guidance in Chapter 1 of DHS Guide to Statistics.
}
\references{
Assaf, S. (2021). Survey_strata.do. GitHub.
https://github.com/DHSProgram/DHS-Analysis-Code/blob/main/Intro_DHSdata_Analysis/2_SurveyDesign/Survey_strata.do

Croft, Trevor N., Allen, Courtney K., Zachary, Blake W., et al. 2023.
Guide to DHS Statistics. Rockville, Maryland, USA: ICF. https://dhsprogram.com/data/Guide-to-DHS-Statistics/index.cfm
}
