% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dhs.R
\name{format_dhs}
\alias{format_dhs}
\title{Benchmark posterior draws via a Metropolis-Hastings algorithm}
\usage{
format_dhs(
  df,
  survey_year,
  period_boundaries,
  strata = c("v023", "v024", "v025")[1]
)
}
\arguments{
\item{df}{a dataframe containing a births recode file from DHS, read into R beforehand using
\code{readstata13::read.dta13}}

\item{survey_year}{the (numeric) year of the DHS survey. If survey spand multiple years,
enter the latest one}

\item{period_boundaries}{a numeric vector containing the years surrounding each time period of interest.
Example: if periods are 2000-2005 and 2005-2010, \code{period_boundaries} should be \code{c(2000,2005,2010)}.}

\item{strata}{a string vector containing which column in \code{df} contains the strata
information. Can be multiple columns. Defaults to "v023".}
}
\value{
A list containing:
\itemize{
\item fitted_list: a list of matrices of benchmarked posterior samples of fitted values in
order arrange(time). Each matrix will have rows arranged in order arrange(region).
\item natl_list: a list of vectors containing aggregated national-level samples that were
accepted, in order arrange(time)
\item prop_accepted: the proportion of samples accepted during sampling.
}
}
\description{
Benchmarks draws from a posterior distribution using a Metropolis-Hastings approach. Takes in
a matrix of posterior draws, a matrix of fixed effect draws, and national-level benchmark data.
}
\author{
Taylor Okonek
}
