% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_turnbull.R
\name{format_turnbull}
\alias{format_turnbull}
\title{Formats output from \code{format_dhs} in appropriate form for \code{turnbull}}
\usage{
format_turnbull(df)
}
\arguments{
\item{df}{output of \code{format_dhs}}
}
\value{
A dataframe containing the births data in a format that can be input to \code{turnbull}.
}
\description{
Formats output from \code{format_dhs} in appropriate way for input to \code{turnbull}. Takes in
the exact output of \code{format_dhs}
}
\author{
Taylor Okonek
}
