% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dhs.R
\name{format_dhs}
\alias{format_dhs}
\title{Formats DHS survey data in appropriate form for parametric survival model for synthetic children.}
\usage{
format_dhs(
  df,
  survey_year,
  period_boundaries,
  strata = c("v023", "v024", "v025")[1]
)
}
\arguments{
\item{df}{a dataframe containing a births recode file from DHS, read into R beforehand using
\code{readstata13::read.dta13}}

\item{survey_year}{the (numeric) year of the DHS survey. If survey spand multiple years,
enter the latest one}

\item{period_boundaries}{a numeric vector containing the years surrounding each time period of interest.
Example: if periods are 2000-2005 and 2005-2010, \code{period_boundaries} should be \code{c(2000,2005,2010)}.}

\item{strata}{a string vector containing which column in \code{df} contains the strata
information. Can be multiple columns. Defaults to "v023".}
}
\value{
A dataframe containing the births recode in a format that can be input to \code{surv_synthetic}.
}
\description{
Formats DHS survey data in appropriate way for input to \code{surv_synthetic}. Takes in
a dataframe containing a births recode file from DHS.
}
\author{
Taylor Okonek
}
