% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_synthetic.R
\name{surv_synthetic}
\alias{surv_synthetic}
\title{Fit a parametric survival model to DHS data across multiple time periods.}
\usage{
surv_synthetic(
  df,
  individual = "individual",
  household = "household",
  cluster = "cluster",
  strata = "strata",
  weights = "weights",
  p = "p",
  a_pi = "a_pi",
  l_p = "l_p",
  I_i = "I_i",
  A_i = "A_i",
  t_i = "t_i",
  t_0i = "t_0i",
  t_1i = "t_1i",
  only_scale = FALSE,
  dist = "weibull"
)
}
\arguments{
\item{df}{a dataframe containing the output from \code{format_dhs}, or optionally, dataframe
containing the following columns}

\item{individual}{column corresponding to individual ID in \code{df}}

\item{household}{column corresponding to household ID in \code{df}}

\item{cluster}{column corresponding to cluster ID in \code{df}}

\item{strata}{column corresponding to strata ID in \code{df}}

\item{weights}{column corresponding to weights in \code{df}}

\item{p}{column corresponding to period ID (numeric) in \code{df}}

\item{a_pi}{column corresponding to child's age at beginning of time period in \code{df}}

\item{l_p}{column corresponding to length of time period in \code{df}}

\item{I_i}{column corresponding to indicator for interval censoring in \code{df}. Should be 1 if
child is interval-censored, 0 if right-censored.}

\item{A_i}{column corresponding to indicator for if a child is interval-censored across the
boundary of a time period in \code{df}}

\item{t_i}{column corresponding to age at right-censoring, if right-censored, in \code{df}}

\item{t_0i}{column corresponding to lower bound of interval, if interval-censored, in \code{df}}

\item{t_1i}{column corresponding to upper bound of interval, if interval-censored, in \code{df}}

\item{only_scale}{boolean for varying only the scale parameter across time period. Defaults to
\code{FALSE}.}

\item{dist}{distribution. Currently only supports "weibull"}
}
\value{
A named listed containing: (1) a dataframe of results, (2) the output from \code{optim},
(3) the gradient evaluated at the MLE, (4) the survey design object, (5) runtime for likelihood
optimization
}
\description{
Fits a parametric survival model to DHS data (formatted from \code{format_dhs}) to get survey-weighted
parameter estimates and finite population variances for parametric survival curves in each time period.
}
\author{
Taylor Okonek
}
