% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_synthetic.R
\name{surv_synthetic}
\alias{surv_synthetic}
\title{Fit a parametric survival model to DHS data across multiple time periods.}
\usage{
surv_synthetic(
  df,
  individual = "individual",
  household = "household",
  cluster = "cluster",
  strata = "strata",
  weights = "weights",
  p = "p",
  a_pi = "a_pi",
  l_p = "l_p",
  I_i = "I_i",
  A_i = "A_i",
  t_i = "t_i",
  t_0i = "t_0i",
  t_1i = "t_1i",
  only_scale = FALSE,
  numerical_grad = FALSE,
  dist = "weibull",
  breakpoints = NA
)
}
\arguments{
\item{df}{a dataframe containing the output from \code{format_dhs}, or optionally, dataframe
containing the following columns}

\item{individual}{column corresponding to individual ID in \code{df}}

\item{household}{column corresponding to household ID in \code{df}}

\item{cluster}{column corresponding to cluster ID in \code{df}}

\item{strata}{column corresponding to strata ID in \code{df}}

\item{weights}{column corresponding to weights in \code{df}}

\item{p}{column corresponding to period ID (numeric) in \code{df}}

\item{a_pi}{column corresponding to child's age at beginning of time period in \code{df}}

\item{l_p}{column corresponding to length of time period in \code{df}}

\item{I_i}{column corresponding to indicator for interval censoring in \code{df}. Should be 1 if
child is interval-censored, 0 if right-censored.}

\item{A_i}{column corresponding to indicator for if a child is interval-censored across the
boundary of a time period in \code{df}}

\item{t_i}{column corresponding to age at right-censoring, if right-censored, in \code{df}}

\item{t_0i}{column corresponding to lower bound of interval, if interval-censored, in \code{df}}

\item{t_1i}{column corresponding to upper bound of interval, if interval-censored, in \code{df}}

\item{only_scale}{boolean for varying only the scale parameter across time period. Defaults to
\code{FALSE}. This option is only available for dist = "weibull"}

\item{numerical_grad}{boolean for whether gradient should be calculated numerically or
analytically. Analytical gradient is faster, but only available for Weibull and Exponential distributions
at the moment.}

\item{dist}{distribution. Currently supports "weibull", "exponential", "piecewise_exponential", "gengamma"}

\item{breakpoints}{if distribution is "piecewise_exponential", the breakpoints (in months) where
the distribution should be divided}
}
\value{
A list containing:
\itemize{
\item result: a dataframe of summarized results
order arrange(time). Each matrix will have rows arranged in order arrange(region).
\item optim: the output from \code{optim}
\item grad: the gradient evaluated at the MLE
\item variance: the finite population variance-covariance matrix
\item design: the survey design object
\item runtime: runtime for likelihood optimization
}
}
\description{
Fits a parametric survival model to DHS data (formatted from \code{format_dhs}) to get survey-weighted
parameter estimates and finite population variances for parametric survival curves in each time period.
}
\details{
We use the original generalized gamma distribution
described in Stacy 1962, where if $\omega \sim Gamma(k, 1)$, then $x = \exp(\omega / shape + \log(scale))$
follows the original generalized gamma distribution. With $shape = b > 0$ and $scale = a > 0$, $x$
has the pdf
$$
f(x \mid a, b, k) = \frac{b}{\Gamma(k)} \frac{x^{bk - 1}}{a^{bk}} \exp(-(x/a)^b)
$$
We note that an alternative parameterisation developed in Prentice (1974) is typically preferred,
as it is more numerically stable in some cases.
}
\author{
Taylor Okonek
}
