% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_synthetic.R
\name{surv_synthetic}
\alias{surv_synthetic}
\title{Fit a parametric survival model to DHS data across multiple time periods.}
\usage{
surv_synthetic(
  df,
  individual = "individual",
  survey = TRUE,
  household = NULL,
  cluster = NULL,
  strata = NULL,
  weights = NULL,
  p = "p",
  a_pi = "a_pi",
  l_p = "l_p",
  I_i = "I_i",
  A_i = "A_i",
  t_i = "t_i",
  t_0i = "t_0i",
  t_1i = "t_1i",
  only_scale = FALSE,
  numerical_grad = FALSE,
  dist = "weibull",
  breakpoints = NA,
  init_vals = NA
)
}
\arguments{
\item{df}{a dataframe containing the output from \code{format_dhs}, or optionally, dataframe
containing the following columns}

\item{individual}{column corresponding to individual ID in \code{df}}

\item{survey}{boolean for whether or not your data comes from a survey with a specified probability design.
If true, must specify household, cluster, strata, and weights parameters in the function, and a finite
population variance based on a pseudo-likelihood will be returned for your survey-weighted estimates.
If false, superpopulation estimates will be returned with the usual asymptotic variance estimator.
Defaults to \code{TRUE}.}

\item{household}{column corresponding to household ID in \code{df}}

\item{cluster}{column corresponding to cluster ID in \code{df}}

\item{strata}{column corresponding to strata ID in \code{df}}

\item{weights}{column corresponding to weights in \code{df}. If survey = FALSE, you may
still specify a weights column, but note that the superpopulation variance will be returned
as opposed to the finite population variance.}

\item{p}{column corresponding to period ID (numeric) in \code{df}}

\item{a_pi}{column corresponding to child's age at beginning of time period in \code{df}}

\item{l_p}{column corresponding to length of time period in \code{df}}

\item{I_i}{column corresponding to indicator for interval censoring in \code{df}. Should be 1 if
child is interval-censored, 0 if right-censored.}

\item{A_i}{column corresponding to indicator for if a child is interval-censored across the
boundary of a time period in \code{df}}

\item{t_i}{column corresponding to age at right-censoring, if right-censored, in \code{df}}

\item{t_0i}{column corresponding to lower bound of interval, if interval-censored, in \code{df}}

\item{t_1i}{column corresponding to upper bound of interval, if interval-censored, in \code{df}}

\item{only_scale}{boolean for varying only the scale parameter across time period. Defaults to
\code{FALSE}. This option is only available for dist = "weibull"}

\item{numerical_grad}{boolean for whether gradient should be calculated numerically or
analytically. Analytical gradient is faster, but only available for Weibull and Exponential distributions
at the moment.}

\item{dist}{distribution. Currently supports "weibull", "exponential",
"piecewise_exponential", "gengamma", "lognormal", "gompertz", "etsp" (exponentially-truncated shifted power family),
"loglogistic"}

\item{breakpoints}{if distribution is "piecewise_exponential", the breakpoints (in months) where
the distribution should be divided}

\item{init_vals}{an optional vector of initial values at which to start the optimizer for the
parameters. Must specify the appropriate number of parameters for the given distribution /
number of periods.}
}
\value{
A list containing:
\itemize{
\item result: a dataframe of summarized results
order arrange(time). Each matrix will have rows arranged in order arrange(region).
\item optim: the output from \code{optim}
\item grad: the gradient evaluated at the MLE
\item variance: the finite population variance-covariance matrix
\item design: the survey design object
\item runtime: runtime for likelihood optimization
\item initial_values: initial values for parameters used in likelihood maximization
}
}
\description{
Fits a parametric survival model to DHS data (formatted from \code{format_dhs}) to get survey-weighted
parameter estimates and finite population variances for parametric survival curves in each time period.
}
\details{
We use the same parameterization of the generalized gamma distribution
as used in \code{\link{flexsurv}}, as opposed to the original parameterization
described in Stacy 1962, where if \eqn{\omega \sim Gamma(k, 1)}, then \eqn{x = \exp(\omega / shape + \log(scale))}
follows the original generalized gamma distribution. With \eqn{shape = b > 0} and \eqn{scale = a > 0}, \eqn{x}
has the pdf

\deqn{f(x \mid a, b, k) = \frac{b}{\Gamma(k)} \frac{x^{bk - 1}}{a^{bk}} \exp(-(x/a)^b)}

The alternative parameterization developed in Prentice (1974) is typically preferred,
as it is more numerically stable in some cases. Under this alternative, if
\eqn{\gamma \sim Gamma(Q^{-2}, 1)} and \eqn{\omega = \log(Q^2 \gamma) / Q}, then \eqn{x = \exp(\mu + \sigma \omega)}
follows the generalized gamma distribution with pdf

\deqn{f(x \mid \mu, \sigma, Q) = \frac{|Q| (Q^{-2}^{Q^{-2}})}{\sigma x \Gamma(Q^{-2})} \exp(Q^{-2}(Q \omega - \exp(Q \omega)))}

The "etsp" distribution uses the exponentially-trunacted shifted power family of hazards defined by

\deqn{h(x) = a(x + c)^{-p} e^{-b x}},

in Scholey (2019), but sets \eqn{c = 0}.
}
\references{
Stacy, E. W. (1962). A generalization of the gamma
distribution.  Annals of Mathematical Statistics 33:1187-92.

Prentice, R. L. (1974). A log gamma model and its maximum likelihood
estimation. Biometrika 61(3):539-544.

Scholey, J (2019). The Age-Trajectory of Infant Mortality in the United States:
Parametric Models and Generative Mechanisms. Annual meeting of the Population
Association of America, Austin, TX. 2019.
}
\author{
Taylor Okonek
}
