% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnbull.R
\name{turnbull}
\alias{turnbull}
\title{Function to obtain Turbull estimates for arbitrarily censored and truncated observations}
\usage{
turnbull(
  df,
  t_0i = "t_0i",
  t_1i = "t_1i",
  lefttrunc = "lefttrunc",
  righttrunc = "righttrunc",
  weights = "weights",
  niter
)
}
\arguments{
\item{df}{a dataframe containing the output from \code{format_dhs}, or optionally, dataframe
containing the following columns}

\item{t_0i}{column corresponding to lower bound of interval if interval-censored,
censoring time if right-censored, observed time if exactly observed, in \code{df}}

\item{t_1i}{column corresponding to upper bound of interval if interval-censored,
\code{Inf} if right-censored, observed time if exactly observed, in \code{df}}

\item{lefttrunc}{column corresponding to left-truncation time in \code{df}, 0 if not
left-truncated}

\item{righttrunc}{column corresponding to right-truncation time \code{df}, \code{Inf} if not
right-truncated}

\item{weights}{column corresponding to weights in \code{df}}

\item{niter}{number of iterations to run the algorithm}
}
\value{
a list containing the following:
\itemize{
\item a dataframe of final estimates for specific times
\item a matrix of estimates at each iteration, with ncol = niter
\item number of iterations
}
}
\description{
Function to obtain Turbull estimates for arbitrarily censored and truncated observations
}
\author{
Taylor Okonek
}
